% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{areOptions}
\alias{areOptions}
\title{Check presence of options in portfolio}
\usage{
areOptions(portfolio)
}
\arguments{
\item{portfolio}{portfolio object}
}
\value{
true if there are, false otherwise
}
\description{
Check presence of options in portfolio
}
