% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{calcExecutionAmts}
\alias{calcExecutionAmts}
\title{Calculate amounts of options execution for given eurusd rate}
\usage{
calcExecutionAmts(portfolio, xRate)
}
\arguments{
\item{portfolio}{portfolio object}

\item{xRate}{eurusd rate at expiration}
}
\value{
vector of execution amounts
}
\description{
Calculate amounts of options execution for given eurusd rate
}
