% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{calcExecutionLosses}
\alias{calcExecutionLosses}
\title{Calculate losses from options execution}
\usage{
calcExecutionLosses(portfolio, xRate)
}
\arguments{
\item{portfolio}{portfolio object}

\item{xRate}{eurusd rate at expiration}
}
\value{
vector of losses
}
\description{
Calculate losses from options execution
}
