% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{calcExpectedFinalWeights}
\alias{calcExpectedFinalWeights}
\title{Calculate expected weight of euro in given portfolio}
\usage{
calcExpectedFinalWeights(portfolio, xRates = portfolio$xRate, avgNum = 100)
}
\arguments{
\item{portfolio}{portfolio object}

\item{xRates}{current eurusd rate}

\item{avgNum}{number of samples in averaging distribution}
}
\value{
expected weight of euro
}
\description{
Calculate expected weight of euro in given portfolio
}
