% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{calcFinalWeights}
\alias{calcFinalWeights}
\title{Calculate final weight of euro}
\usage{
calcFinalWeights(portfolio, finalXRates)
}
\arguments{
\item{portfolio}{portfolio object}

\item{finalXRates}{final eurusd rate}
}
\value{
final euro weight
}
\description{
Calculate final weight of euro
}
