% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{calcRebalanceAmt}
\alias{calcRebalanceAmt}
\title{Calculate euro amount for rebalancing trade toward expected weight if with options,
toward target weight otherwise}
\usage{
calcRebalanceAmt(portfolio, newXRate, avgNum = 100)
}
\arguments{
\item{portfolio}{portfolio object}

\item{newXRate}{new eurusd rate}

\item{avgNum}{number of samples in averaging distribution}
}
\value{
euro amount of rebalancing trade
}
\description{
Calculate euro amount for rebalancing trade toward expected weight if with options,
toward target weight otherwise
}
