% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{calcRFactor}
\alias{calcRFactor}
\title{Calculate R-factor}
\usage{
calcRFactor(portfolio, newXRate, avgNum = 100)
}
\arguments{
\item{portfolio}{portfolio object}

\item{newXRate}{new eurusd rate}

\item{avgNum}{number of samples in averaging distribution}
}
\value{
R-factor
}
\description{
Calculate R-factor
}
