% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{chartSpreadLossDistributions}
\alias{chartSpreadLossDistributions}
\title{Chart spread loss distributions}
\usage{
chartSpreadLossDistributions(portfolio, sampSize = 10^2)
}
\arguments{
\item{portfolio}{portfolio object}

\item{sampSize}{sample size}
}
\value{
charts of spread loss distributions for no option and with option rebalancing
}
\description{
Chart spread loss distributions
}
