% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{chartWeightDistributions}
\alias{chartWeightDistributions}
\title{Chart final eur weight distributions for no rebalancing and rebalancing with options only}
\usage{
chartWeightDistributions(portfolio)
}
\arguments{
\item{portfolio}{portfolio object}
}
\value{
charts of final eur weight distributions
}
\description{
Chart final eur weight distributions for no rebalancing and rebalancing with options only
}
