% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{delOptions}
\alias{delOptions}
\title{Delete options from portfolio}
\usage{
delOptions(portfolio)
}
\arguments{
\item{portfolio}{portfolio object}
}
\value{
portfolio object without options
}
\description{
Delete options from portfolio
}
