% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{demoSeekingBestOptions}
\alias{demoSeekingBestOptions}
\title{Demonstrate seeking of best options}
\usage{
demoSeekingBestOptions()
}
\value{
best options
}
\description{
Demonstrate seeking of best options
}
