% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{findMaxSpreadLossPath}
\alias{findMaxSpreadLossPath}
\title{Find example of path with high total spread loss}
\usage{
findMaxSpreadLossPath(portfolio, sampSize = 10^3, avgNum = 100)
}
\arguments{
\item{portfolio}{portfolio object}

\item{sampSize}{sample size}

\item{avgNum}{number of samples in averaging distribution}
}
\value{
vector of eurusd path
}
\description{
Find example of path with high total spread loss
}
