% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{getRebalanceAnalytics}
\alias{getRebalanceAnalytics}
\title{Get rebalance analytics}
\usage{
getRebalanceAnalytics(portfolio, path, avgNum = 100,
  getTradeRate = getTradeRateDefault)
}
\arguments{
\item{portfolio}{portfolio object}

\item{path}{vector of eurusd path rates}

\item{avgNum}{number of samples in averaging distribution}

\item{getTradeRate}{unction that returns trade rate from market rate and amount}
}
\value{
data.frame with data of trade amounts, spread losses and weights
}
\description{
Get rebalance analytics
}
