% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{getTradeRateDefault}
\alias{getTradeRateDefault}
\title{Get trade eurusd rate with spread loss
Default function}
\usage{
getTradeRateDefault(xRate, amount)
}
\arguments{
\item{xRate}{fair market eurusd rate}

\item{amount}{amount of trade}
}
\value{
eurusd trade rate
}
\description{
Get trade eurusd rate with spread loss
Default function
}
