% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{getValue}
\alias{getValue}
\title{Get portfolio value in usd}
\usage{
getValue(portfolio)
}
\arguments{
\item{portfolio}{portfolio object}
}
\value{
portfolio value in usd
}
\description{
Get portfolio value in usd
}
