% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{getValuesOfPortfolioOptions}
\alias{getValuesOfPortfolioOptions}
\title{Get values of given portfolio options}
\usage{
getValuesOfPortfolioOptions(portfolio, xRate = portfolio$xRate,
  vols = numeric(length(portfolio$options$strikes)) + portfolio$volaty)
}
\arguments{
\item{portfolio}{portfolio object}

\item{xRate}{current eurusd rate}

\item{vols}{volatility of eurusd rate}
}
\value{
values of options in usd
}
\description{
Get values of given portfolio options
}
