% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{getWeight}
\alias{getWeight}
\title{Get euro weight in portfolio}
\usage{
getWeight(portfolio)
}
\arguments{
\item{portfolio}{portfolio object}
}
\value{
euro weight
}
\description{
Get euro weight in portfolio
}
