% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{Portfolio}
\alias{Portfolio}
\title{Constructor of portfolio object}
\usage{
Portfolio()
}
\value{
Portfolio object
}
\description{
Attributes:
$eur - amount of euro in eur
$usd - amount of usd
$options$amounts - amounts of options in eur, positive for put, negative for call
$options$strikes - options strikes
$wTarget - target weight of euro
$xRate - current eurusd rate
$iRateUsd - usd interest rate
$iRateEur - eur interest rate
$volaty - volatility of eurusd rate
$daysToEnd - days to the end of period, expiration of options
$daysInYear - number of business days in year
$spreadLoss - current accumulated spread loss form FX spot rebalancing trades
}
