% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{rebalance}
\alias{rebalance}
\title{Rebalance portfolio}
\usage{
rebalance(portfolio, newXRate, avgNum = 100,
  getTradeRate = getTradeRateDefault)
}
\arguments{
\item{portfolio}{portfolio object}

\item{newXRate}{new eurusd rate}

\item{avgNum}{number of samples in averaging distribution}

\item{getTradeRate}{function that returns trade rate from market rate and amount}
}
\value{
rebalanced portfolio object
}
\description{
Rebalance portfolio
}
