% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{rebalanceAlongPath}
\alias{rebalanceAlongPath}
\title{Rebalance portfolio along given path}
\usage{
rebalanceAlongPath(portfolio, path, avgNum = 100)
}
\arguments{
\item{portfolio}{portfolio object}

\item{path}{vector of eurusd path rates}

\item{avgNum}{number of samples in averaging distribution}
}
\value{
rebalanced portfolio object
}
\description{
Rebalance portfolio along given path
}
