% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{simFinResult}
\alias{simFinResult}
\title{Simulate financial results}
\usage{
simFinResult(portfolio, optTradeVols, sampSize = 10^3, avgNum = 100)
}
\arguments{
\item{portfolio}{portfolio object}

\item{optTradeVols}{vector of options trade volatilities}

\item{sampSize}{sample size}

\item{avgNum}{number of samples in averaging distribution}
}
\value{
data frame of following samples:
 no options spread loss, with options spread loss, optrions execution loss,
  premium received, financial result
}
\description{
Simulate financial results
}
