% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{simOptionsLoss}
\alias{simOptionsLoss}
\title{Simulate option losses}
\usage{
simOptionsLoss(portfolio, sampSize = 10^3)
}
\arguments{
\item{portfolio}{portfolio object}

\item{sampSize}{sample size}
}
\value{
vector of option losses
}
\description{
Simulate option losses
}
