% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{simSpreadLoss}
\alias{simSpreadLoss}
\title{Simulate total spread losses accumulated during the period}
\usage{
simSpreadLoss(portfolio, sampSize = 10^3, avgNum = 100)
}
\arguments{
\item{portfolio}{portfolio object}

\item{sampSize}{sample size}

\item{avgNum}{number of samples in averaging distribution}
}
\value{
vector of spread loss samples
}
\description{
Simulate total spread losses accumulated during the period
}
