% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{simXPath}
\alias{simXPath}
\title{Simulate random walk path of eurusd rate for portfolio$daysToEnd days}
\usage{
simXPath(portfolio)
}
\arguments{
\item{portfolio}{portfolio object}
}
\value{
vector of eurusd rate path
}
\description{
Simulate random walk path of eurusd rate for portfolio$daysToEnd days
}
