% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxor.R
\name{simXRates}
\alias{simXRates}
\title{Simulate eurusd rate in portfolio$daysToEnd days}
\usage{
simXRates(portfolio, xRate = portfolio$xRate, sampSize = 10^3,
  isRandom = FALSE)
}
\arguments{
\item{portfolio}{portfolio object}

\item{xRate}{start eurusd rate}

\item{sampSize}{sample size}

\item{isRandom}{are samples random normal or evenly normal distributed}
}
\value{
vector of eurusd rate samples
}
\description{
Simulate eurusd rate in portfolio$daysToEnd days
}
